Finalize new component detection format
