+++ resources/projects/open-liberty/src/main/liberty/config/server.xml
@@ -11,11 +11,7 @@
 
   <mpMetrics authentication="false" />
 
-  <variable name="default.http.port" defaultValue="9080"/>
-  <variable name="default.https.port" defaultValue="9443"/>
-  <variable name="app.context.root" defaultValue="/"/>
-
-  <httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="${default.https.port}" id="defaultHttpEndpoint"/>
+  <httpEndpoint host="*" httpPort="9080" httpsPort="9443" id="defaultHttpEndpoint"/>
 
   <webApplication location="io.openliberty.sample.getting.started.war" contextRoot="${app.context.root}"/>
 </server>
