+++ resources/projects/spring/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+server.http.port=9012
